Move ServerTest to spec/support
no functional changes, just to make integration test shorter
